add csrf token and csrf validation

+++ src/App/Controller/DeleteController.php
@@ -6,6 +6,9 @@ use Gasparik\App\Flash;
 use Gasparik\Lib\Application\Controller;
 use Gasparik\Lib\Request\Request;
 use Gasparik\Lib\Response\Response;
+use Gasparik\Lib\Validation\CsrfRule;
+use Gasparik\Lib\Validation\RequiredRule;
+use Gasparik\Lib\Validation\Validation;
 use Gasparik\Lib\Websupport\WebsupportApi;
 
 class DeleteController implements Controller
@@ -19,6 +22,14 @@ class DeleteController implements Controller
 
     public function execute(Request $request): Response
     {
+        $validator = Validation::makeWithCsrf([]);
+        $errors = $validator->validate($request->getInputList());
+        if (!empty($errors)) {
+            $errroKey = array_key_first($errors);
+            Flash::error($errors[$errroKey]);
+            return Response::redirect('/');
+        }
+
         $response = $this->websupportApi->deleteDnsRecord($request->getInput('id', 0));
         if (!isset($response['status']) || $response['status'] !== 'success') {
             // todo: improve error message. Use response error message if it provides some additional information for user
